CORE-11198 only remove CPIs if locally not in a Jenkins env

+++ testing/e2e-test-utilities/src/main/kotlin/net/corda/e2etest/utilities/CpiLoader.kt
@@ -81,7 +81,10 @@ object CpiLoader {
             // Read CPI
             return cpiPath.readAll().inputStream()
         } finally {
-            tempDirectory.deleteRecursively()
+            // only remove in a local environment, not in Jenkins where JENKINS_URL is always set
+            if (System.getenv("JENKINS_URL") == null) {
+                tempDirectory.deleteRecursively()
+            }
         }
     }
